Major refactor
This does a couple things. Primarily, it introduces connection pooling
to work around the issue that Kafka can only process a single request at
a time per TCP connection. Given that many requests might be waiting for
production/consumption timeouts, there's little reason to block other
client requests on those timeouts.

By switching to using a connection pool, we can detect when all
connections are blocked and attempt to be more intelligent. There are a
number of tunables for helping prevent stampeding and other issues.

This also reworks most of the mutex usage including moving metadata
out to its own type and file. This should help reason about things and
reduce the complexity a bit in the broker file.