Clean up connection leaks in broker and offset coordinator
When fetching a new connection for the offset coordinator, the broker
would leak the connection used for the metadata request. Moreover, there
was no way to reclaim the connection used by an offset coordinator. We
cleaned up the connection leak and added a Close() method for the
coordinator which returns its connection to the broker's pool.