Add debugging to ConnectionPool
This makes the connection pool more aware of failing to get a connection
due to "too many" and adds debugging of existing connections around
that. This also changes a bit of the way it works to be slightly less
precise at the risk it might use an extra connection or two in a racy
situation which is better than not having connections available,
usually.